Properly separate data of different Abteilungen
Each sensitive data point is marked with the id of the sync group
(Abteilung for now, maybe canton in the future). Also, the id column in
the tables is no longer misused to store the id from MiData, because we
potentially need to store the same entity twice for different sync
groups (e.g. when the same person is in multiple Abteilungen).